feat : HUD 화면 비율 설정

+++ Assets/Scripts/SaveData.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public class SaveData
 {
-    public int currentScore;
+    public long currentScore;
     public string lastSaveTime; // 마지막 끌떄 시간
 }
